Resolve merge conflicts with develop branch
Agent-Logs-Url: https://github.com/ROCm/rocm-systems/sessions/8b512c91-ffa5-4cb3-a3e0-602d16fa1cca

+++ projects/rocshmem/src/bootstrap/bootstrap.cpp
@@ -68,7 +68,7 @@ struct ExtInfo {
    int rank_pos = -1;
 
    // Confirm that rank is in the vectors of ranks
-   for (int i = 0; i < ranks.size(); i++) {
+   for (size_t i = 0; i < ranks.size(); i++) {
      if (rank == ranks[i]) {
        rank_pos = i;
        break;
@@ -104,7 +104,7 @@ struct ExtInfo {
    int rank_pos = -1;
 
    // Confirm that rank is in the vectors of ranks
-   for (int i = 0; i < ranks.size(); i++) {
+   for (size_t i = 0; i < ranks.size(); i++) {
      if (rank == ranks[i]) {
        rank_pos = i;
        break;
@@ -122,7 +122,7 @@ struct ExtInfo {
    char *recv_buf = new char[size * num_pes];
    std::memset(recv_buf, 0, num_pes * size);
 
-   // Perform pairwise exchange - local copy is ommitted
+   // Perform pairwise exchange - local copy is omitted
    for (int step = 1; step < num_pes; step++) {
      int sendto   = (rank_pos + step) % num_pes;
      int recvfrom = (rank_pos + num_pes - step) % num_pes;
